test: Use sigjmp_buf instead of jmp_buf for the sigjmp() function.
This change is made to be conformant with the IEEE Std 1003.1-2001,
also known as POSIX 2001, which explicitly requires the function to
be called with struct sigjmp_buf as the first parameter.

While glibc defines both data-structures to be identical, other
libc implementations may define them individually, and thus, from
the POV of the C compiler to be different and/or incompatible types.